project/services/weather/pyproject.toml
[tool.poetry]
name = "weather"
version = "0.1.0"
description = ""
authors = ["mirek <mirek@cnl.sk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = "^0.26.0"
fastapi-restful = "^0.5.0"
typing-inspect = "^0.9.0"
paho-mqtt = "^1.6.1"
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
